pyproject.toml
[tool.poetry]
name = "intercompy"
version = "1.1.0"
description = "Headless intercom using Raspberry Pi GPIO and Telegram"
authors = ["John Casey <jdcasey@commonjava.org>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
"ruamel.yaml" = "^0.17.21"
click = "^8.1.3"
"RPi.GPIO" = "^0.7.1"
Pyrogram = "^2.0.41"
TgCrypto = "^1.2.3"
PyAudio = "^0.2.12"
ffmpy = "^0.3.0"
gTTS = "^2.2.4"
pyttsx3 = "^2.90"
python-vlc = "^3.0.16120"
SpeechRecognition-ForkedVersion = "^3.9.2"
pydub = "^0.25.1"
nltk = "^3.7"
opentelemetry-api = "^1.12.0"
opentelemetry-sdk = "^1.12.0"
opentelemetry-exporter-otlp-proto-http = "^1.12.0"
opentelemetry-instrumentation-requests = "^0.33b0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[tool.scmver]
enable = true
write_to = "version.py"

[tool.poetry.scripts]
intercom = "intercompy.command:run"
intercompy-test-gpio = "intercompy.command:selftest_gpio"
intercompy-session-setup = "intercompy.command:session_setup"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
